refactor(widget): #24 Signup screen indentation
